Move calculations to c file
